Finished player movement
1. Fixed ceiling interaction and a bug (where the player would freeze
when the edge of the player perfectly landed on the edge of a platform)
by making sure the player is not touching a wall or ground and then if
the player has less than a certain velocity, move forward.

2. Made actioncount, the variable that keeps track of double jumping,
update in LateUpdate so that the player doesnt refresh his jumps after
jumping resulting in 1 extra jump than intended.

3.  Fixed dash so that you can jump directly out of dash and hitting any
object stops your dash.  I set the counter to equal dash time so that
the while loop ends and the coroutine ends. When dash is pressed, the
counter is set to 0.